+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/Chargeable.java
@@ -2,8 +2,8 @@ package kenakata.catalog;
 
 public interface Chargeable {
 
-double unitVat();
-double unitCharge();
+long unitVat();
+long unitCharge();
 String label();
 
 }
